+++ RDS.ExpenseTracker.Business/Models/FinancialAccount.cs
@@ -5,7 +5,7 @@ namespace RDS.ExpenseTracker.Business.Models
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public decimal Availability { get; set; }
+        public int Availability { get; set; }
         public string? Description { get; set; }
     }
 }
