+++ RDS.ExpenseTrackerApi/Dtos/FinancialAccountDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public double? Availability { get; set; }
+        public int? Availability { get; set; }
         public string? Description { get; set; }
     }
 }
